requirements.txt
# GPU碰撞检测系统依赖
# Python 3.8+

# GPU加速计算
cupy-cuda11x>=12.0.0  # 使用CUDA 11.x版本，根据您的CUDA版本选择

# 科学计算
numpy>=1.21.0
scipy>=1.7.0

# 可视化
matplotlib>=3.5.0
opencv-python>=4.5.0

# 性能分析（可选）
line-profiler>=3.5.0
memory-profiler>=0.60.0

# 开发工具（可选）
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
